Update README.md
Celisus/Fahrenheit input validation

Examine the code and run the program. Try entering -500 as an input for Fahrenheit. You'll notice that the Fahrenheit prompt has input validation to prevent inputs below absolute zero (the lowest limit on the temperature scale).

Modify the code so that the prompt for Celsius also has input validation. Absolute zero in Celsius is -273.15